style: run linters

+++ contracts/Interfaces/Compound/UniswapAnchoredViewI.sol
@@ -4,13 +4,13 @@ pragma solidity 0.6.12;
 interface UniswapAnchoredViewI {
     /**
      * @notice Get the underlying price of a cToken
-     * @dev Implements the PriceOracle interface for Compound v2. 
+     * @dev Implements the PriceOracle interface for Compound v2.
      * See: https://docs.compound.finance/v2/prices/#underlying-price
      * @param cToken The cToken address for price retrieval
-     * @return The price of the asset in USD as an unsigned integer scaled up by 10 ^ (36 - underlying asset decimals). 
+     * @return The price of the asset in USD as an unsigned integer scaled up by 10 ^ (36 - underlying asset decimals).
      * E.g. WBTC has 8 decimal places, so the return value is scaled up by 1e28.
      */
-    function getUnderlyingPrice(address cToken) external view returns (uint);
+    function getUnderlyingPrice(address cToken) external view returns (uint256);
 
     /**
      * @notice Get the official price for a symbol
@@ -18,5 +18,5 @@ interface UniswapAnchoredViewI {
      * @param symbol The symbol to fetch the price of
      * @return The price of the asset in USD as an unsigned integer scaled up by 10 ^ 6.
      */
-    function price(string memory symbol) external view returns (uint);
+    function price(string memory symbol) external view returns (uint256);
 }
